Move Okio utils tests to a separate suite
